backend/requirements.txt
fastapi
uvicorn
fastrtc
pydantic
openai
black
pyjwt
python-multipart
python-jose[cryptography]